requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.3
pymongo==4.5.0
google-generativeai==0.3.0
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
SpeechRecognition==3.10.0
moviepy==1.0.3
pydub==0.25.1
weasyprint==60.2
reportlab==4.0.4

# Optional dependencies for enhanced functionality
# For better PDF extraction
pdfplumber==0.9.0

# For better video transcription (optional - requires significant space)
# openai-whisper==20231117

# Development dependencies
pytest==7.4.2
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0
gunicorn